Completely rework devinabox to be easier for core devs to use.
Previously devinabox was essentially a script for people to use to
pull together a directory full of the code needed to contribute to
Python. Unfortunately the script turned out to be brittle, requiring
patches for every sprint it was used with.

Since the users of this project have continually been core devs,
re-orient towards core devs leading sprints. This means a README
outlining what hg repos to clone, a script to help generate a coverage
report using coverage.py, and a script for new contributors to use to
help build CPython quickly and with sane defaults in a single step.
All other details should be gleaned from the devguide.